better merge_default_subquery_branch
